\name{e_dfmvclass}
\alias{e_dfmvclass}
\title{e_dfmvclass}
\usage{
e_dfmvclass()
}
\description{
Explaining datasets
}
\examples{

  arrest
  coffee
  heywood
  KR_authority
  KR_senior_center
  OMKthesis
  Path_Ingram
  smartphone2018
  YS_authority
  YS_commitment
  YS_jobsatisfaction
  YS_jobsatisfaction2

}
