\name{e_dfmvclass}
\alias{e_dfmvclass}
\title{e_dfmvclass}
\usage{
e_dfmvclass()
}
\description{
Explaining datasets
}
\examples{

  #-----------------------------
  arrest
  coffee
  heywood
  OMKthesis
  Path_Ingram
  smartphone2018

  # -----------------------------
  KR_authority           #김렬
  KR_senior_center       

  # -----------------------------
  L_authority           #이석환
  L_commitment          
  L_jobsatisfaction     
  L_jobsatisfaction2    

  #-----------------------------
  base                   #이석민
  basic                  
  con_moderator1
  con_moderator2
  fat
  fat2
  group_moderator
  iris_missing
  lgcm
  matched
  mediation
  mg_sem1
  mg_sem2
  mg_sem2_L
  mroz
  psm_sem

  # -----------------------------
  ch10                   #김계수
  ch3
  ch31
  ch4
  ch9
  ch9_1
  data
  education
  education_100
  ex10
  ex31
  ex42
  exdata
}
