\name{e_dfmvclass}
\alias{e_dfmvclass}
\title{e_dfmvclass}
\usage{
e_dfmvclass()
}
\description{
Explaining datasets
}
\examples{

  #-----------------------------
  arrest
  coffee
  heywood
  OMKthesis
  Path_Ingram
  smartphone2018

  #-----------------------------
  KR_authority           #김O
  KR_senior_center       #O렬
  YS_authority           #이OO
  YS_commitment          #O석O
  YS_jobsatisfaction     #이OO
  YS_jobsatisfaction2    #이OO

  #-----------------------------
  base                   #이O민
  basic                  #이석O
  con_moderator1
  con_moderator2
  fat
  fat2
  group_moderator
  iris_missing
  lgcm
  matched
  mediation
  mg_sem1
  mg_sem2
  mg_sem2_L
  mroz
  psm_sem

}
